Examples of combinations of options
The backgrounds with the important spells for the game (having spells that don't appear on the Sorcerer or Warlock spell list are the ones from Guildmaster's Guide to Ravnica. The ones that offer healing are the Boris Legionaire (Aid + paladin-esque spells), and the Selesnya Initiate (Aid+Conjure Minor Elementals). If you take a background with a healing spell, take a race that does not offer healing spells in order to maximize the breadth of your abilities.
The races with important spells for the game are: Halfling (Mark of Hospitality) [healing via Aid+Leomund's Tiny Hut+Create Food and Water+Aura of Purity+Private Sanctum+Hallow, ...]; Human (Mark of Handling) [mostly important for the powerful conjuration spell Conjure Animals]; Halfling (Mark of Healing) [all the important healing spells between 1st and 5th level, making for a very powerful Sorcerer with the ability to cast Mass Healing Word or Healing Word as a reaction using only 2 additional Sorcery points]; Dwarf (Mark of Warding) (Alarm, a free cast of Arcane Lock (no material components), Mordenkainen's Secret Chest and Faithful Hound), Human (Mark of Making) (Fabricate, Stone Shape, Creation, etc.).